Made progress on Rating system
Started work on the forms for rating a tutor, due to me not knowing how to automatically input a tutor based of the zoom link, I have created a form to select a tutor from a list. I also began work on the front end for the rating page. and made necessary migrations for the review system to work (I hope).